requirements.txt
faster-whisper==1.0.2
groq==0.6.0
openai==1.30.1
google-generativeai==0.5.4
opencv-python==4.9.0.80
Pillow==10.3.0
pyaudio==0.2.14
SpeechRecognition==3.10.1
pyperclip==1.8.2
requests==2.31.0
gtts==2.3.2
pydub==0.25.1
pyttsx3==2.90
python-dotenv==1.0.0
